build NNF in a separate orchestrator
